Added ScrollAreas For Tasks And Scripts
Although I have added the scroll areas, I can't scroll xD. I have solved another issue that kept happening, the app opens up not in the maximized state (despite writing self.showMaximized () in the constructor), and when maximized it would mess with the grid layout. My solution was when updating to check on the state of the window, and if changed it would reinitialise the UI again, which fixes the grid layout.